BUG/MEDIUM: server/ssl: Don't reuse TLS session if the connection's SNI differs
When a new SSL server connection is created, if no SNI is set, it is
possible to inherit from the one of the reused TLS session. The bug was
introduced by the commit 95ac5fe4a ("MEDIUM: ssl_sock: always use the SSL's
server name, not the one from the tid"). The mixup is possible between
regular connections but also with health-checks connections.

But it is only the visible part of the bug. If the SNI of the cached TLS
session does not match the one of the new connection, no reuse must be
performed at all.

To fix the bug, the connection SNI is now retrieved earlier, to be available
when the SSL context is initialized and when we try to reuse a TLS
session. So before calling conn_prepare() function. This SNI is temporarily
saved in the connection itself to be compared in
ssl_sock_srv_try_reuse_sess() function.

This patch should fix the issue #3195. It must be slowly backported as far
as 3.0.

On 3.3, it relies on:

  * MINOR: connection: Add a SNI member in the connection structure
  * MEDIUM: quic: Add connection as argument when qc_new_conn() is called

On older versions, only the first one must be backported but this patch must
also be adapted. On these versions, ssl_sock_srv_try_reuse_sess() does not
exist and changes must be applied on ssl_sock_init().

WARNING: Because of this patch, the TLS session reuse no longer works on
         WolfSSL if a SNI is used on the connection because the SNI cannot
         be retrieved. It is the reason why the reuse is not performed. But
         on older versions, the SNI is not available too and this remains an
         issue too.