REGTESTS: ssl: No longer test SSL reuse with WolfSSL
When fixing the issue #3195, it appeared that the SNI cannot be retrived
from server connection with WolfSSL. It is of course an issue. But since the
commit "BUG/MEDIUM: server/ssl: Don't reuse TLS session if the connection's
SNI differs", it also means the SSL reuse no longer work with WolfSSL.

It is clearly an issue with WolfSSL (or wit the way we are using it, I
honestly don't know). So let's skip the regtests on this feature when
HAProxy is complied with WolfSSL.