BUG/MEDIUM: ssl-sock: Never reset the TLS session when the SNI is set
This revert the commit 119a4084b ("BUG/MEDIUM: ssl: for a handshake when
server-side SNI changes"). Thanks to the previous patch, the SNI of the now
connection always matches the one of the reused TLS session.

And it is especially important because without this fix, but with the
previous one, the resume no longer works (because the previous SNI is not
set anymore when the xprt is initialized).

In addition, nothing is performed if there is no SNI. It is useless for most
libraries but it is mandatory for WolfSSL because it crashes if the SNI is
NULL.

This patch must be bacported with the following one, as far as 3.0:

  * BUG/MEDIUM: server/ssl: Don't reuse TLS session if the connection's SNI differs