MINOR: ha-inject: Split the C object files list for ha-inject
Add new OBJS_COMMON makefile variable to list the C object files
which are common to haproxy and new ha-inject tool.
Add ha-inject new target to build this tool. It depends on OJBS_HA_INJECT
variable which must list the C oject files specific to ha-inject in
addition to OBJS_COMMON list.
This is a first split to be completed. Note that ha-inject does not
already compile due to the parsing part which has been excluded from
OBJS_COMMON. Indeed, a lot of C file depends on cfgparse.c functions
to register their own keywords.

Add ha-inject.c new C file. It is very basic to define the main() function
for ha-inject.